Закончил с бронированием

+++ src/modules/reservations/reservations.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { Model, Types } from 'mongoose'
 import {
@@ -7,25 +7,59 @@ import {
   ReservationSearchOptions
 } from './interfaces/reservation.interface'
 import { Reservation, ReservationDocument } from './schemas/reservation.schema'
+import { HttpException } from '@nestjs/common'
+//import { User } from 'src/common/decorators/my-custom.decorator'
 
 @Injectable()
 export class ReservationsService implements IReservation {
   constructor(
     @InjectModel(Reservation.name)
-    private ReservationModel: Model<ReservationDocument>
+    private ReservationModel: Model<ReservationDocument> //@Inject(User) private user
   ) {}
 
-  addReservation(data: ReservationDto): Promise<ReservationDocument> {
+  async addReservation(data: ReservationDto): Promise<ReservationDocument> {
     //TODO: Метод IReservation.addReservation должен проверять, доступен ли номер на заданную дату.
+    const reservations = await this.ReservationModel.find({ room: data.room })
+    if (reservations) {
+      if (
+        reservations.find(
+          (res: ReservationDocument) =>
+            (new Date(data.dateStart) >= new Date(res.dateStart) &&
+              new Date(data.dateStart) <= new Date(res.dateEnd)) ||
+            (new Date(data.dateEnd) >= new Date(res.dateStart) &&
+              new Date(data.dateEnd) <= new Date(res.dateEnd))
+        )
+      )
+        return null
+    }
     return new this.ReservationModel(data).save()
   }
 
-  removeReservation(id: Types.ObjectId): Promise<void> {
-    this.ReservationModel.findByIdAndRemove(id).exec()
-    return
+  async getReservations(
+    filter: ReservationSearchOptions
+  ): Promise<Reservation[]> {
+    return await this.ReservationModel.find(filter, {
+      user: 0,
+      _id: 0,
+      __v: 0
+    })
+      .populate({
+        path: 'hotel',
+        select: { title: 1, description: 1, _id: 0 }
+      })
+      .populate({
+        path: 'room',
+        select: { images: 1, description: 1, _id: 0 }
+      })
+      .exec()
   }
 
-  getReservations(filter: ReservationSearchOptions): Promise<Reservation[]> {
-    return this.ReservationModel.find(filter).exec()
+  async removeReservation(id: Types.ObjectId): Promise<void> {
+    const reservation = await this.ReservationModel.findById(id)
+    if (!reservation) throw new HttpException('Резервация не найдена', 400)
+    //TODO:как получить данные пользователя в сервисе?
+    //if (reservation.user != this.user.id) throw new HttpException('Доступ запрещен', 403)
+    this.ReservationModel.findByIdAndRemove(id).exec()
+    return
   }
 }
